requirements.txt
discord.py

youtube_dl

requests

psutil

PyNaCl

ffmpeg


ROLES:
- Countdown Master
- Admin
- Captain